Rename util files. There is no point having _util postfix when already in utils/ pkg path.
