+++ auto_gptq_next/utils/__init__.py
@@ -1,1 +1,1 @@
-from .perplexity_utils import Perplexity
+from .perplexity import Perplexity
