+++ tests/test_q4_cuda.py
@@ -1,7 +1,7 @@
 import unittest  # noqa: E402
 
 import torch  # noqa: E402
-from auto_gptq_next.utils.import_utils import dynamically_import_QuantLinear  # noqa: E402
+from auto_gptq_next.utils.importer import dynamically_import_QuantLinear  # noqa: E402
 from parameterized import parameterized  # noqa: E402
 
 try:
